requirements.txt
pytz
psycopg2 ~= 2.6.0
pillow ~= 3.3.1
xlrd ~= 1.0.0
jsonschema ~= 2.5.1
WeasyPrint ~= 0.31.0
pytils ~= 0.3.0
pysberbps ~= 0.0.1.dev1
dicttoxml ~= 1.7.4

pytest ~= 2.9.2
pytest-django ~= 2.9.1
factory-boy ~= 2.7.0
pytest-bdd ~= 2.18.1

django ~= 1.10.0
django-bulk-update ~=1.1.0
django-rq ~=0.9.2
django-admin-decorators ~= 0.1.0

django-webpack-loader ~= 0.3.3

djangorestframework ~= 3.4.0
django-rest-swagger ~= 0.3.0
djangorestframework_camel_case ~= 0.2.0
django-filter ~= 0.14.0
drf-nested-routers ~=0.10.0

requests ~= 2.11.0
django-ipware ~= 1.1.0
beautifulsoup4 ~= 4.5.0
